fix: generalize streamed tool argument parse errors
