+++ src_py/agenthub/deepseek_v4/__init__.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .client import DeepSeekV4Client, ToolCallArgumentParseError
+from .client import DeepSeekV4Client
 
 
-__all__ = ["DeepSeekV4Client", "ToolCallArgumentParseError"]
+__all__ = ["DeepSeekV4Client"]
