+++ src_ts/src/kimi_k2_6/client.ts
@@ -20,6 +20,7 @@ import type {
   ChatCompletionCreateParamsStreaming,
 } from "openai/resources/chat/completions";
 import { LLMClient } from "../baseClient";
+import { parseToolCallArguments } from "../errors";
 import {
   EventType,
   FinishReason,
@@ -460,7 +461,12 @@ export class KimiK2_6Client extends LLMClient {
                   {
                     type: "tool_call",
                     name: partialToolCall.name,
-                    arguments: JSON.parse(partialToolCall.arguments || "{}"),
+                    arguments: parseToolCallArguments(
+                      partialToolCall.arguments,
+                      "kimi_k2_6",
+                      partialToolCall.name || "",
+                      partialToolCall.tool_call_id || "",
+                    ),
                     tool_call_id: partialToolCall.tool_call_id || "",
                   },
                 ],
@@ -489,7 +495,12 @@ export class KimiK2_6Client extends LLMClient {
               {
                 type: "tool_call",
                 name: partialToolCall.name,
-                arguments: JSON.parse(partialToolCall.arguments || "{}"),
+                arguments: parseToolCallArguments(
+                  partialToolCall.arguments,
+                  "kimi_k2_6",
+                  partialToolCall.name || "",
+                  partialToolCall.tool_call_id || "",
+                ),
                 tool_call_id: partialToolCall.tool_call_id || "",
               },
             ],
